[ast] Always use 32 bit integers

+++ samlang-core-services/__tests__/source-processor.test.ts
@@ -25,19 +25,19 @@ it('hello world processor test', () => {
   expect(compileTimeErrors.map((it) => it.toString())).toEqual([]);
 
   const llvmModule = lowerSourcesToLLVMModules(checkedSources).forceGet(moduleReference);
-  expect(prettyPrintLLVMModule(llvmModule)).toBe(`declare i64* @_builtin_malloc(i64) nounwind
-declare i64 @_builtin_println(i64*) nounwind
-declare i64 @_builtin_throw(i64*) nounwind
-declare i64* @_builtin_intToString(i64) nounwind
-declare i64 @_builtin_stringToInt(i64*) nounwind
-declare i64* @_builtin_stringConcat(i64*, i64*) nounwind
+  expect(prettyPrintLLVMModule(llvmModule)).toBe(`declare i32* @_builtin_malloc(i32) nounwind
+declare i32 @_builtin_println(i32*) nounwind
+declare i32 @_builtin_throw(i32*) nounwind
+declare i32* @_builtin_intToString(i32) nounwind
+declare i32 @_builtin_stringToInt(i32*) nounwind
+declare i32* @_builtin_stringConcat(i32*, i32*) nounwind
 
 ; @GLOBAL_STRING_0 = 'Hello World!'
-@GLOBAL_STRING_0 = private unnamed_addr constant [13 x i64] [i64 12, i64 72, i64 101, i64 108, i64 108, i64 111, i64 32, i64 87, i64 111, i64 114, i64 108, i64 100, i64 33], align 8
-define i64 @_compiled_program_main() local_unnamed_addr nounwind {
+@GLOBAL_STRING_0 = private unnamed_addr constant [13 x i32] [i32 12, i32 72, i32 101, i32 108, i32 108, i32 111, i32 32, i32 87, i32 111, i32 114, i32 108, i32 100, i32 33], align 8
+define i32 @_compiled_program_main() local_unnamed_addr nounwind {
 l0_start:
-  %_temp_0_string_name_cast = bitcast [13 x i64]* @GLOBAL_STRING_0 to i64*
-  call i64 @_builtin_println(i64* %_temp_0_string_name_cast) nounwind
-  ret i64 0
+  %_temp_0_string_name_cast = bitcast [13 x i32]* @GLOBAL_STRING_0 to i32*
+  call i32 @_builtin_println(i32* %_temp_0_string_name_cast) nounwind
+  ret i32 0
 }`);
 });
